ZSWS revisions
This commit implements a number of changes to the SWS implementation on the micro:bit to account for nonsense characters.

In a future commit, this will be re-engineered to use a user provided buffer for reception. This is to reduce the overhead of maintaining a circular buffer and to simplify code.