lowered timer priority and fixed race.
After lowering the timer priority, there was a race condition where if a UART interrupt occurred during a timer interrupt and event was created, it would corrupt the timestamp of the timer.

For safety we just disable all interrupts. I guess it can be removed if the protocol can run at a lower priority.